Further changes

+++ run.sh
@@ -64,7 +64,9 @@ REPOS_CSV_WITH_HASHES="${REPOS_CSV::length-4}_with_hashes.csv"
 mkdir -p "$OUT_DIR"
 
 # Delete all locks in cache
-find "$CACHE_DIR" -name "*.lock" -delete
+if [ -d "$CACHE_DIR" ]; then
+    find "$CACHE_DIR" -name "*.lock" -delete
+fi
 
 python3 src/python/write_head_hashes.py \
     --repos_csv "$REPOS_CSV" \
@@ -95,16 +97,15 @@ python3 src/python/merge_filter.py \
 python3 src/python/merge_tester.py \
     --valid_repos_csv "$OUT_DIR/valid_repos.csv" \
     --merges_path "$OUT_DIR/merges_analyze/" \
-    --output_dir "$OUT_DIR" \
+    --output_dir "$OUT_DIR/merges_tested/" \
     --cache_dir "$CACHE_DIR/test_results"
 
-exit 0
-
 python3 src/python/latex_output.py \
+    --tested_merges_path "$OUT_DIR/merges_tested/" \
     --full_repos_csv "$REPOS_CSV" \
     --valid_repos_csv "$OUT_DIR/valid_repos.csv" \
     --n_merges "$N_MERGES" \
     --result_csv "$OUT_DIR/result.csv" \
-    --merges_path "$OUT_DIR/merges/" \
+    --all_merges_path "$OUT_DIR/merges/" \
     --merges_valid_path "$OUT_DIR/merges_valid/" \
-    --output_path "$OUT_DIR"
+    --output_dir "$OUT_DIR"
